forensics-pdf-mcp/requirements.txt
fastmcp>=2.2.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
PyMuPDF>=1.24.0
Pillow>=10.0.0
httpx>=0.27.0
cryptography>=42.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
respx>=0.21.0
